refactor retry functinoality into a separate package
