+++ aliases_test.go
@@ -27,7 +27,7 @@ func TestAliasIntegrationSuite(t *testing.T) {
 func (s *AliasIntegrationSuite) TestGetAlias() {
 	ctx := test.Context()
 
-	alias, err := s.rc.GetAlias(ctx, persistentWorkspace, persistentAlias)
+	alias, err := s.rc.GetAlias(ctx, test.PersistentWorkspace, test.PersistentAlias)
 	require.NoError(s.T(), err)
 	assert.Equal(s.T(), "pme@rockset.com", alias.GetCreatorEmail())
 }
@@ -45,7 +45,7 @@ func (s *AliasIntegrationSuite) TestListAliases() {
 func (s *AliasIntegrationSuite) TestListAliasesForWorkspace() {
 	ctx := test.Context()
 
-	aliases, err := s.rc.ListAliases(ctx, option.WithAliasWorkspace(persistentWorkspace))
+	aliases, err := s.rc.ListAliases(ctx, option.WithAliasWorkspace(test.PersistentWorkspace))
 	require.NoError(s.T(), err)
 	for _, a := range aliases {
 		s.T().Logf("workspace: %s", a.GetName())
@@ -55,17 +55,17 @@ func (s *AliasIntegrationSuite) TestListAliasesForWorkspace() {
 func (s *AliasIntegrationSuite) TestAliases() {
 	ctx := test.Context()
 
-	_, err := s.rc.CreateAlias(ctx, persistentWorkspace, s.alias, []string{"commons._events"})
+	_, err := s.rc.CreateAlias(ctx, test.PersistentWorkspace, s.alias, []string{"commons._events"})
 	require.NoError(s.T(), err)
 
-	err = s.rc.WaitUntilAliasAvailable(ctx, persistentWorkspace, s.alias)
+	err = s.rc.WaitUntilAliasAvailable(ctx, test.PersistentWorkspace, s.alias)
 	require.NoError(s.T(), err)
 
 	// update
 
-	err = s.rc.WaitUntilAliasAvailable(ctx, persistentWorkspace, s.alias)
+	err = s.rc.WaitUntilAliasAvailable(ctx, test.PersistentWorkspace, s.alias)
 	require.NoError(s.T(), err)
 
-	err = s.rc.DeleteAlias(ctx, persistentWorkspace, s.alias)
+	err = s.rc.DeleteAlias(ctx, test.PersistentWorkspace, s.alias)
 	require.NoError(s.T(), err)
 }
