+++ internal/test/client.go
@@ -13,3 +13,9 @@ func Client(t *testing.T) *rockset.RockClient {
 	require.NoError(t, err)
 	return rc
 }
+
+func DebugClient(t *testing.T) *rockset.RockClient {
+	rc, err := rockset.NewClient(rockset.WithUserAgent("rockset-go-integration-tests"), rockset.WithHTTPDebug())
+	require.NoError(t, err)
+	return rc
+}
